requirements.txt
PyYAML>=3.12
numpy>=1.15.0
tensorflow>=1.9.0
tensorboard>=1.9.0
scikit-learn>=0.19.0
